requirements.txt
emoji
rich
nltk
openpyxl
python-docx
PyPDF2
pandas
streamlit
plotly
